add readwrite, update security tests
